chore: optimized code
