+++ package.json
@@ -66,7 +66,6 @@
     "preinstall": "npx only-allow pnpm",
     "setup": "rm -rf node_modules packages/*/node_modules packages/*/lib packages/*/es packages/*/dist && pnpm install",
     "build": "node scripts/build",
-    "build:diff": "node scripts/build-diff",
     "lint:fix": "eslint --ext .js,.jsx,.ts,.tsx packages --fix",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx packages",
     "check-and-publish": "node scripts/check-and-publish",
